+++ taxonomy/urls.py
@@ -1,0 +1,8 @@
+"""
+Taxonomy v1 API URLs.
+"""
+from django.urls import include, path
+
+urlpatterns = [
+    path('api/', include('taxonomy.api.urls')),
+]